+++ Common/src/main/java/com/nyfaria/spookybats/client/renderer/VoidBatRenderer.java
@@ -1,5 +1,6 @@
 package com.nyfaria.spookybats.client.renderer;
 
+import com.mojang.blaze3d.vertex.PoseStack;
 import com.nyfaria.spookybats.client.model.SpookyBatModel;
 import com.nyfaria.spookybats.client.renderer.api.SpookyBatRenderer;
 import com.nyfaria.spookybats.entity.api.SpookyBat;
@@ -11,6 +12,11 @@ public class VoidBatRenderer extends SpookyBatRenderer<SpookyBat> {
 
     public VoidBatRenderer(EntityRendererProvider.Context p_173929_, SpookyBatModel<SpookyBat> model) {
         super(p_173929_, model);
+    }
+
+    @Override
+    protected void scale(SpookyBat pLivingEntity, PoseStack pMatrixStack, float pPartialTickTime) {
+        super.scale(pLivingEntity, pMatrixStack, pPartialTickTime);
     }
 
     @Nullable