+++ Common/src/main/java/com/nyfaria/spookybats/entity/ai/TeleportAwayGoal.java
@@ -3,6 +3,7 @@ package com.nyfaria.spookybats.entity.ai;
 import com.nyfaria.spookybats.entity.api.SpookyBat;
 import net.minecraft.core.particles.ParticleTypes;
 import net.minecraft.world.entity.ai.goal.Goal;
+import net.minecraft.world.entity.ai.util.AirRandomPos;
 import net.minecraft.world.phys.Vec3;
 
 public class TeleportAwayGoal extends Goal {
@@ -24,7 +25,6 @@ public class TeleportAwayGoal extends Goal {
 		int randX = mob.getRandom().nextInt(7);
 		int randY = mob.getRandom().nextInt(7);
 		int randZ = mob.getRandom().nextInt(7);
-
 		double destX = mob.getRandom().nextInt(2) == 0
 			? batPos.x + randX
 			: batPos.x - randX;
@@ -34,15 +34,6 @@ public class TeleportAwayGoal extends Goal {
 		double destZ = mob.getRandom().nextInt(2) == 0
 			? batPos.z + randZ
 			: batPos.z - randZ;
-
-		mob.level().addParticle(
-			ParticleTypes.PORTAL,
-			mob.getX(),
-			mob.getRandomY(),
-			mob.getZ(),
-			0, 0, 0
-		);
-
-		mob.teleportTo(destX, destY, destZ);
+		mob.randomTeleport(destX, destY, destZ, true);
 	}
 }